+++ flow-typed/platform.js
@@ -40,6 +40,7 @@ declare type HapiServer = {
 		[string]: any,
 	},
 	route: (routes: HapiRoute | Array<HapiRoute>) => Promise<HapiServer>,
+	on: (eventName: string, () => void) => void,
 };
 declare type HapiRequest = {
 	getLanguage: () => string,
@@ -116,21 +117,31 @@ declare type QueryResponse = {
 	flags?: Array<string>,
 	meta?: Object,
 	query?: Query,
+	error?: string,
 };
 
 declare type QueryFunction = (location: { [string]: mixed }) => Query;
 
-declare type PlatformRoute = {
-	component: React$ComponentType<*>,
-	getNestedRoutes?: () => Promise<Array<PlatformRoute>>,
-	getIndexRoute?: () => Promise<PlatformRoute>,
+type BasePlatformRoute = {|
 	path?: string,
 	exact?: boolean,
 	strict?: boolean,
 	query?: QueryFunction | Array<QueryFunction>,
 	indexRoute?: PlatformRoute,
 	routes?: Array<PlatformRoute>,
-};
+|};
+
+type AsyncPlatformRoute = {|
+	...BasePlatformRoute,
+	getComponent: () => Promise<React$ComponentType<*>>,
+|};
+
+type StaticPlatformRoute = {|
+	...BasePlatformRoute,
+	component: React$ComponentType<*>,
+|};
+
+declare type PlatformRoute = AsyncPlatformRoute | StaticPlatformRoute;
 
 declare type CookieOpts = {
 	path?: string,