+++ packages/mwp-api-state/src/reducer.js
@@ -118,6 +118,7 @@ export function app(
 		type: string,
 		payload: { responses: Array<Object> },
 		error?: Error,
+		meta?: any,
 	}
 ) {
 	let newState;